refactor(audit): own command history projection
